Allow `--at boot` as a special start value
I often find myself wanting to start tracking work when I boot up my
computer. However, sometimes I have to rush into a meeting or do
something else that causes me to not immediately start a time frame.

In such a case, I usually look up the boot time later and use that to
start the tracking at the correct time.

This patch allows for the special value `boot` to be used when starting
a frame like this:

    watson start --at boot ...